Add Policies for Comment and Category

+++ app/Http/Controllers/Api/CommentController.php
@@ -14,31 +14,35 @@ class CommentController extends Controller
     public function index():JsonResponse
     {
         $comments = CommentResource::collection(Comment::all());
+        $this->authorize('viewAny', $comments);
 
         return response()->json([
            'comments'=>$comments
         ]);
     }
-    public function show($id): JsonResponse
+    public function show($id, Comment $comment): JsonResponse
     {
-        $comment = CommentResource::make(Comment::find($id));
+        $this->authorize('view', $comment);
+        $comment = CommentResource::make($comment::find($id));
 
         return response()->json([
            'comment'=>$comment
         ]);
     }
-    public function store(CommentRequest $request): JsonResponse
+    public function store(CommentRequest $request, Comment $comment): JsonResponse
     {
-        $comment = Comment::create($request->all());
+        $this->authorize('create', $comment);
+        $comment = $comment::create($request->all());
         $comment = CommentResource::make($comment);
 
         return response()->json([
            'comment'=>$comment
         ]);
     }
-    public function update($id, CommentRequest $request): JsonResponse
+    public function update($id, CommentRequest $request, Comment $comment): JsonResponse
     {
-        $comment = Comment::find($id);
+        $this->authorize('update', $comment);
+        $comment = $comment::find($id);
         $comment->update($request->all());
         $comment->save();
         $comment = CommentResource::make($comment);
@@ -47,8 +51,9 @@ class CommentController extends Controller
            'comment'=>$comment
         ]);
     }
-    public function destroy($id): JsonResponse
+    public function destroy($id, Comment $comment): JsonResponse
     {
+        $this->authorize('delete', $comment);
         $comment = Comment::find($id);
         $comment->delete();
